+++ Venmo/src/main/java/com/braintreepayments/api/VenmoPaymentAuthRequestCallback.java
@@ -6,7 +6,7 @@ import androidx.annotation.Nullable;
  * Callback to handle result from
  * {@link VenmoClient#tokenize(VenmoPaymentAuthResult, VenmoResultCallback)}
  */
-public interface VenmoCreatePaymentAuthRequestCallback {
+public interface VenmoPaymentAuthRequestCallback {
     void onPaymentAuthRequest(@Nullable VenmoPaymentAuthRequest venmoPaymentAuthRequest,
                               @Nullable Exception error);